% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNC.R
\name{baseflow}
\alias{baseflow}
\title{Applies a baseflow estimation filter using Stan}
\usage{
baseflow(
  formula = formula,
  data = data,
  method = "lh",
  passes = 3,
  bfi = 0.8,
  time_method = time_method,
  iter = 3000,
  warmup = 1500,
  chains = 3,
  refresh = max(iter/100, 1),
  seed = seed
)
}
\arguments{
\item{formula}{A formula as in lm()}

\item{data}{A long data frame containing the dates, covariates and the response variable.}

\item{method}{A string specifying the type of filter to use.}

\item{passes}{The number of passes of the filter. Usually 3.}

\item{bfi}{A baseflow index used in the Eckhardt's approach}

\item{time_method}{Defines the type of time series model to capture temporal dependence.}

\item{iter}{Number of iterations}

\item{warmup}{Warm up samples}

\item{chains}{Number of chains}

\item{refresh}{Sampler refreshing rate}

\item{seed}{(optional) A seed for reproducibility}
}
\value{
A list with the model fit

It returns a stanfit object. It includes the formula used to fit the model.
}
\description{
Missing values are allowed in the response but not in the covariates.
}
\details{
Missing values are not allowed in the covariates and they must be imputed before using the function.
}
\examples{
\dontrun{
#options(mc.cores = parallel::detectCores())
# Lyne and Hollick filter with 3 passes
fit_lh <- baseflow(formula = y ~ prec,
                   data = df,
                   method = 'lh',
                  passes = 3)

#Eckhardt approach
fit_eck <- baseflow(formula = y ~ prec,
                    data = df,
                    method = 'eck')
}

}
\author{
Edgar Santos-Fernandez
}
