% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_from_dgp.R
\name{dgp_example_normal}
\alias{dgp_example_normal}
\title{Example data-generating process (normal distribution)}
\usage{
dgp_example_normal(rep_id, n_obs, mu, sigma, ...)
}
\arguments{
\item{rep_id}{id of dataset}

\item{n_obs}{number of observations}

\item{mu}{mu parameter}

\item{sigma}{sigma parameter}

\item{...}{Unused}
}
\value{
List of true parameters (mu and sigma) and resulting observations
}
\description{
Example data-generating process (normal distribution)
}
