% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elpd.R
\name{elpd}
\alias{elpd}
\alias{elpd.stanreg_list}
\title{Expected pointwise predictive density}
\usage{
\method{elpd}{stanreg_list}(x, ...)

elpd.stanreg_list(model_list, weights, ...)
}
\arguments{
\item{x}{List of fitted models.}

\item{...}{arguments passed to `log_lik`, e.g. `newdata`}

\item{weights}{vector of model weights}
}
\value{
`loo` `elpd_generic` object
}
\description{
Compute elpd for a set of models and model weights.
}
