% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_from_dgp.R
\name{generate_from_dgp}
\alias{generate_from_dgp}
\title{Generate datasets from a data-generating process}
\usage{
generate_from_dgp(dgp, n_datasets, ...)
}
\arguments{
\item{dgp}{Data-generating process function}

\item{n_datasets}{Number of datasets to generate}

\item{...}{Arguments passed to data-generating process function}
}
\value{
List of datasets, each generated from the dgp function
}
\description{
Generate datasets from a data-generating process
}
\examples{
\dontrun{
generate_from_dgp(
  dgp = dgp_example_normal,
  n_datasets = 10,
  mean = 5,
  sd = 2
)
}
}
