% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tikz_plot.R
\name{save_tikz_plot}
\alias{save_tikz_plot}
\title{Save a graphic or plot as a tikz file.}
\usage{
save_tikz_plot(plot, filename, width = NA, height = NA, asp = NA)
}
\arguments{
\item{plot}{the object to be saved.}

\item{filename}{A character string indicating the desired path to the output
file. If both arguments are used in the function call, file will be
preferred.}

\item{width}{The width of the output figure, in **inches**.}

\item{height}{The height of the output figure, in **inches**.}

\item{asp}{The aspect ratio of the saved figure.}
}
\value{
save_tikz_plot returns no values.
}
\description{
Save a graphic or plot as a tikz file.
}
