% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_from_dgp.R
\name{dgp_example_normal}
\alias{dgp_example_normal}
\title{Example data-generating process (normal distribution)}
\usage{
dgp_example_normal(n, mean, sd, ...)
}
\arguments{
\item{n}{Number of observations}

\item{mean}{Mean of normal distribution}

\item{sd}{Standard deviation of normal distribution}

\item{...}{unused but required to exist for use with `generate_from_dgp`}
}
\value{
List of true parameters (mu and sigma) and observations
}
\description{
Example data-generating process (normal distribution)
}
